capture mutations without oncokb annotations available
